Move addon settings JS to a separate module
