+++ webpack.common.config.js
@@ -123,6 +123,7 @@ var resolve = {
         'addonHelper': staticPath('js/addonHelper.js'),
         'koHelpers': staticPath('js/koHelpers.js'),
         'addonPermissions': staticPath('js/addonPermissions.js'),
+        'addonSettings': staticPath('js/addonSettings.js'),
         'navbar-control': staticPath('js/navbarControl.js'),
         'markdown': staticPath('js/markdown.js'),
         'diffTool': staticPath('js/diffTool.js'),
